feat: remove excessive description from "Add a secret" dialog
note: the description is an accessibility requirement for screen
readers and our tests will fail without it, because radix ui issues
console warnings for it, and we fail our tests on console warnings.

this behavior makes sense, so instead of disabling it, I decided to
keep the description for screen readers (hence the sr-only class).

In order to make the behavior less confusing when refactoring/debugging
this area of the app, I also added a test case that verifies the
screen-reader behavior.